Move the code from the old file to the new files
